+++ src/Components/InvoicePayment/InvoicePaymentSubmitButton.js
@@ -1,0 +1,6 @@
+import React from "react";
+import { SubmitPaymentMethod } from "../PaymentMethod/SubmitPaymentMethod";
+
+export const invoicePaymentSubmitButton = (props) => {
+  return <SubmitPaymentMethod {...props} />;
+};