+++ src/Components/PaymentMethod/PaymentMethodContainer.js
@@ -45,6 +45,7 @@ import {
 } from "../common/Helpers";
 import {
   Subscription,
+  StripeGateway,
   PaypalGateway,
   VantivGateway,
   SUBSCRIPTION_TYPES
@@ -118,6 +119,7 @@ const PaymentMethodContainerWithoutStripe = ({
     props.giftRecipient ?? pelcroStore.giftRecipient;
   const isRenewingGift =
     props.isRenewingGift ?? pelcroStore.isRenewingGift;
+  const invoice = props.invoice ?? pelcroStore.invoice;
 
   useEffect(() => {
     if (window.Pelcro.coupon.getFromUrl()) {
@@ -725,6 +727,37 @@ const PaymentMethodContainerWithoutStripe = ({
     );
   };
 
+  const payInvoice = (gatewayService, gatewayToken, dispatch) => {
+    const subscription = new Subscription(gatewayService);
+
+    return subscription.execute(
+      {
+        type: SUBSCRIPTION_TYPES.PAY_INVOICE,
+        token: gatewayToken,
+        invoiceId: invoice.id
+      },
+      (err, res) => {
+        dispatch({ type: DISABLE_SUBMIT, payload: false });
+        dispatch({ type: LOADING, payload: false });
+
+        if (err) {
+          onFailure(err);
+          return dispatch({
+            type: SHOW_ALERT,
+            payload: {
+              type: "error",
+              content: getErrorMessages(err)
+            }
+          });
+        }
+
+        set({ invoice: null });
+
+        onSuccess(res);
+      }
+    );
+  };
+
   const updatePaymentSource = (state, dispatch) => {
     return stripe
       .createSource({ type: "card" })
@@ -790,6 +823,10 @@ const PaymentMethodContainerWithoutStripe = ({
         }
 
         const getOrderItemsTotal = () => {
+          if (!order) {
+            return null;
+          }
+
           const isQuickPurchase = !Array.isArray(order);
 
           if (isQuickPurchase) {
@@ -806,7 +843,10 @@ const PaymentMethodContainerWithoutStripe = ({
         };
 
         const totalAmount =
-          state?.updatedPrice ?? plan?.amount ?? getOrderItemsTotal();
+          state?.updatedPrice ??
+          plan?.amount ??
+          invoice.amount_remaining ??
+          getOrderItemsTotal();
 
         if (
           source?.card?.three_d_secure === "required" &&
@@ -838,6 +878,10 @@ const PaymentMethodContainerWithoutStripe = ({
    * @return {Promise}
    */
   const resolveTaxCalculation = () => {
+    if (type === "invoicePayment") {
+      return new Promise((resolve) => resolve());
+    }
+
     const taxesEnabled = window.Pelcro.site.read()?.taxes_enabled;
 
     return new Promise((resolve, reject) => {
@@ -922,6 +966,8 @@ const PaymentMethodContainerWithoutStripe = ({
       subscribe(stripeSource, state, dispatch);
     } else if (stripeSource && type === "orderCreate") {
       purchase(stripeSource, state, dispatch);
+    } else if (stripeSource && type === "invoicePayment") {
+      payInvoice(new StripeGateway(), stripeSource.id, dispatch);
     }
   };
 
@@ -1072,9 +1118,17 @@ const PaymentMethodContainerWithoutStripe = ({
           );
 
         case HANDLE_PAYPAL_SUBSCRIPTION:
-          return UpdateWithSideEffect(state, () =>
-            handlePaypalSubscription(state, action.payload)
-          );
+          return UpdateWithSideEffect(state, (state, dispatch) => {
+            if (type === "invoicePayment") {
+              payInvoice(
+                new PaypalGateWay(),
+                action.payload,
+                dispatch
+              );
+            } else {
+              handlePaypalSubscription(state, action.payload);
+            }
+          });
 
         case SET_UPDATED_PRICE:
           return Update({ ...state, updatedPrice: action.payload });