+++ src/components.js
@@ -186,6 +186,17 @@ export { EmailVerifyModal } from "./Components/EmailVerify/EmailVerifyModal";
 export { EmailVerifyResendButton } from "./Components/EmailVerify/EmailVerifyResendButton";
 export { EmailVerifyView } from "./Components/EmailVerify/EmailVerifyView";
 
+export { InvoicePaymentContainer } from "./Components/InvoicePayment/InvoicePaymentContainer";
+export { InvoicePaymentModal } from "./Components/InvoicePayment/InvoicePaymentModal";
+export { invoicePaymentSubmitButton } from "./Components/InvoicePayment/InvoicePaymentSubmitButton";
+export { InvoicePaymentView } from "./Components/InvoicePayment/InvoicePaymentView";
+
+export { InvoiceDetailsContainer } from "./Components/InvoiceDetails/InvoiceDetailsContainer";
+export { InvoiceDetailsModal } from "./Components/InvoiceDetails/InvoiceDetailsModal";
+export { InvoiceDetailsPayButton } from "./Components/InvoiceDetails/InvoiceDetailsPayButton";
+export { InvoiceDetailsDownloadButton } from "./Components/InvoiceDetails/InvoiceDetailsDownloadButton";
+export { InvoiceDetailsView } from "./Components/InvoiceDetails/InvoiceDetailsView";
+
 export { AlertWithContext as Alert } from "./SubComponents/AlertWithContext";
 export { Alert as AlertElement } from "./SubComponents/Alert";
 export { Badge } from "./SubComponents/Badge";