+++ src/services/Subscription/Subscription.service.js
@@ -12,7 +12,8 @@ export const SUBSCRIPTION_TYPES = {
   CREATE_SUBSCRIPTION: "CREATE_SUBSCRIPTION",
   CREATE_GIFTED_SUBSCRIPTION: "CREATE_GIFTED_SUBSCRIPTION",
   RENEW_SUBSCRIPTION: "RENEW_SUBSCRIPTION",
-  RENEW_GIFTED_SUBSCRIPTION: "RENEW_GIFTED_SUBSCRIPTION"
+  RENEW_GIFTED_SUBSCRIPTION: "RENEW_GIFTED_SUBSCRIPTION",
+  PAY_INVOICE: "PAY_INVOICE"
 };
 
 export class Subscription {
@@ -40,6 +41,7 @@ export class Subscription {
    * @property {number} [subscriptionIdToRenew]
    * @property {number} [quantity]
    * @property {string} addressId
+   * @property {number} invoiceId
    */
 
   /**
@@ -133,6 +135,8 @@ export class StripeGateway {
         return this.#createGiftedSubscription(options, callback);
       case types.RENEW_GIFTED_SUBSCRIPTION:
         return this.#renewGiftedSubscription(options, callback);
+      case types.PAY_INVOICE:
+        return this.#payInvoice(options, callback);
 
       default:
         console.error(
@@ -272,6 +276,21 @@ export class StripeGateway {
       }
     );
   };
+
+  #payInvoice = (options, callback) => {
+    const { token, invoiceId } = options;
+
+    window.Pelcro.invoice.pay(
+      {
+        payment_gateway: this.#paymentGateway,
+        gateway_token: token,
+        invoice_id: invoiceId
+      },
+      (err, res) => {
+        callback(err, res);
+      }
+    );
+  };
 }
 
 /**
@@ -294,6 +313,8 @@ export class PaypalGateway {
         return this.#createSubscription(options, callback);
       case types.CREATE_GIFTED_SUBSCRIPTION:
         return this.#createGiftedSubscription(options, callback);
+      case types.PAY_INVOICE:
+        return this.#payInvoice(options, callback);
 
       default:
         console.error(
@@ -365,6 +386,21 @@ export class PaypalGateway {
         gift_start_date: giftRecipient?.startDate,
         gift_message: giftRecipient?.giftMessage,
         address_id: product.address_required ? addressId : null
+      },
+      (err, res) => {
+        callback(err, res);
+      }
+    );
+  };
+
+  #payInvoice = (options, callback) => {
+    const { token, invoiceId } = options;
+
+    window.Pelcro.invoice.pay(
+      {
+        payment_gateway: this.#paymentGateway,
+        gateway_token: token,
+        invoice_id: invoiceId
       },
       (err, res) => {
         callback(err, res);