+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+# [3.0.0-beta.7](https://github.com/pelcro-inc/react-pelcro-js/compare/v3.0.0-beta.6...v3.0.0-beta.7) (2022-03-03)
+
+
+### Bug Fixes
+
+* **donations menu:** showing all expired subs ([0477f6c](https://github.com/pelcro-inc/react-pelcro-js/commit/0477f6c87386469dcd770fed5343d566962e945f))
+
+# [3.0.0-beta.6](https://github.com/pelcro-inc/react-pelcro-js/compare/v3.0.0-beta.5...v3.0.0-beta.6) (2022-03-02)
+
+
+### Bug Fixes
+
+* **(url triggers):** invoice details trigger not working ([72446f9](https://github.com/pelcro-inc/react-pelcro-js/commit/72446f97945ccde4f203fe3b3d1f228c9c7e842b))
+* **donations menu:** not showing canceled donations ([03d0bae](https://github.com/pelcro-inc/react-pelcro-js/commit/03d0bae5ce751747f77130747964356c752c1dd3))
+* **url trigger:** invalid invoice id error msg wording ([b206ad8](https://github.com/pelcro-inc/react-pelcro-js/commit/b206ad837f37042b65f06d3ff9a07c3212ebc116))
+
+# [3.0.0-beta.5](https://github.com/pelcro-inc/react-pelcro-js/compare/v3.0.0-beta.4...v3.0.0-beta.5) (2022-03-02)
+
+
+### Bug Fixes
+
+* **(url triggers):** invoice details trigger not working ([c8c74a0](https://github.com/pelcro-inc/react-pelcro-js/commit/c8c74a0b83095a5ffb12533d37b8f12a5b3a8696))
+
+# [3.0.0-beta.4](https://github.com/pelcro-inc/react-pelcro-js/compare/v3.0.0-beta.3...v3.0.0-beta.4) (2022-03-01)
+
+
+### Features
+
+* add ability to view and pay invoices ([#133](https://github.com/pelcro-inc/react-pelcro-js/issues/133)) ([a5f95f3](https://github.com/pelcro-inc/react-pelcro-js/commit/a5f95f39955da0863af1e00aa5e697f98155ce1a))
+
 # [3.0.0-beta.3](https://github.com/pelcro-inc/react-pelcro-js/compare/v3.0.0-beta.2...v3.0.0-beta.3) (2022-02-23)